Move environment to hGraphQL
